+++ index.html
@@ -6,7 +6,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -165,6 +164,6 @@
       </div>
     </section>
   </template>
-
+  <script type="module" src="./js/main.js" ></script>
 </body>
 </html>
